pyproject.toml
[project]
name = "chat-rag"
version = "0.1.0"
description = "Ask question about your documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.17",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.5",
    "langchain-huggingface>=0.1.2",
    "langchain>=0.3.7",
    "pymupdf>=1.24.13",
    "sentence-transformers>=3.2.1",
    "streamlit>=1.39.0",
]

[dependency-groups]
dev = [
    "ruff>=0.7.2",
]
